+++ src/js/internal/validators.ts
@@ -12,7 +12,7 @@ const tokenRegExp = /^[\^_`a-zA-Z\-0-9!#$%&'*+.|~]+$/;
  * per the rules defined in RFC 7230
  * See https://tools.ietf.org/html/rfc7230#section-3.2.6
  */
-function checkIsHttpToken(val) {
+export function checkIsHttpToken(val) {
   return RegExpPrototypeExec.$call(tokenRegExp, val) !== null;
 }
 
@@ -35,7 +35,7 @@ function validateLinkHeaderFormat(value, name) {
   }
 }
 
-function validateLinkHeaderValue(hints) {
+export function validateLinkHeaderValue(hints) {
   if (typeof hints === "string") {
     validateLinkHeaderFormat(hints, "hints");
     return hints;
@@ -68,45 +68,43 @@ function validateLinkHeaderValue(hints) {
 }
 hideFromStack(validateLinkHeaderValue);
 
-export default {
+
   /** (value, name) */
-  validateObject: $newCppFunction("NodeValidator.cpp", "jsFunction_validateObject", 2),
-  validateLinkHeaderValue: validateLinkHeaderValue,
-  checkIsHttpToken: checkIsHttpToken,
+ export const validateObject = $newCppFunction("NodeValidator.cpp", "jsFunction_validateObject", 2);
   /** `(value, name, min, max)` */
-  validateInteger: $newCppFunction("NodeValidator.cpp", "jsFunction_validateInteger", 0),
+ export const validateInteger = $newCppFunction("NodeValidator.cpp", "jsFunction_validateInteger", 0);
   /** `(value, name, min, max)` */
-  validateNumber: $newCppFunction("NodeValidator.cpp", "jsFunction_validateNumber", 0),
+ export const validateNumber = $newCppFunction("NodeValidator.cpp", "jsFunction_validateNumber", 0);
   /** `(value, name)` */
-  validateString: $newCppFunction("NodeValidator.cpp", "jsFunction_validateString", 0),
+ export const validateString = $newCppFunction("NodeValidator.cpp", "jsFunction_validateString", 0);
   /** `(number, name)` */
-  validateFiniteNumber: $newCppFunction("NodeValidator.cpp", "jsFunction_validateFiniteNumber", 0),
+ export const validateFiniteNumber = $newCppFunction("NodeValidator.cpp", "jsFunction_validateFiniteNumber", 0);
   /** `(number, name, lower, upper, def)` */
-  checkRangesOrGetDefault: $newCppFunction("NodeValidator.cpp", "jsFunction_checkRangesOrGetDefault", 0),
+ export const checkRangesOrGetDefault = $newCppFunction("NodeValidator.cpp", "jsFunction_checkRangesOrGetDefault", 0);
   /** `(value, name)` */
-  validateFunction: $newCppFunction("NodeValidator.cpp", "jsFunction_validateFunction", 0),
+ export const validateFunction = $newCppFunction("NodeValidator.cpp", "jsFunction_validateFunction", 0);
   /** `(value, name)` */
-  validateBoolean: $newCppFunction("NodeValidator.cpp", "jsFunction_validateBoolean", 0),
+ export const validateBoolean = $newCppFunction("NodeValidator.cpp", "jsFunction_validateBoolean", 0);
   /** `(port, name = 'Port', allowZero = true)` */
-  validatePort: $newCppFunction("NodeValidator.cpp", "jsFunction_validatePort", 0),
+ export const validatePort = $newCppFunction("NodeValidator.cpp", "jsFunction_validatePort", 0);
   /** `(signal, name)` */
-  validateAbortSignal: $newCppFunction("NodeValidator.cpp", "jsFunction_validateAbortSignal", 0),
+ export const validateAbortSignal = $newCppFunction("NodeValidator.cpp", "jsFunction_validateAbortSignal", 0);
   /** `(value, name, minLength = 0)` */
-  validateArray: $newCppFunction("NodeValidator.cpp", "jsFunction_validateArray", 0),
+ export const validateArray = $newCppFunction("NodeValidator.cpp", "jsFunction_validateArray", 0);
   /** `(value, name, min = -2147483648, max = 2147483647)` */
-  validateInt32: $newCppFunction("NodeValidator.cpp", "jsFunction_validateInt32", 0),
+ export const validateInt32 = $newCppFunction("NodeValidator.cpp", "jsFunction_validateInt32", 0);
   /** `(value, name, positive = false)` */
-  validateUint32: $newCppFunction("NodeValidator.cpp", "jsFunction_validateUint32", 0),
+ export const validateUint32 = $newCppFunction("NodeValidator.cpp", "jsFunction_validateUint32", 0);
   /** `(signal, name = 'signal')` */
-  validateSignalName: $newCppFunction("NodeValidator.cpp", "jsFunction_validateSignalName", 0),
+ export const validateSignalName = $newCppFunction("NodeValidator.cpp", "jsFunction_validateSignalName", 0);
   /** `(data, encoding)` */
-  validateEncoding: $newCppFunction("NodeValidator.cpp", "jsFunction_validateEncoding", 0),
+ export const validateEncoding = $newCppFunction("NodeValidator.cpp", "jsFunction_validateEncoding", 0);
   /** `(value, name)` */
-  validatePlainFunction: $newCppFunction("NodeValidator.cpp", "jsFunction_validatePlainFunction", 0),
+ export const validatePlainFunction = $newCppFunction("NodeValidator.cpp", "jsFunction_validatePlainFunction", 0);
   /** `(value, name)` */
-  validateUndefined: $newCppFunction("NodeValidator.cpp", "jsFunction_validateUndefined", 0),
+ export const validateUndefined = $newCppFunction("NodeValidator.cpp", "jsFunction_validateUndefined", 0);
   /** `(buffer, name = 'buffer')` */
-  validateBuffer: $newCppFunction("NodeValidator.cpp", "jsFunction_validateBuffer", 0),
+ export const validateBuffer = $newCppFunction("NodeValidator.cpp", "jsFunction_validateBuffer", 0);
   /** `(value, name, oneOf)` */
-  validateOneOf: $newCppFunction("NodeValidator.cpp", "jsFunction_validateOneOf", 0),
-};
+ export const validateOneOf = $newCppFunction("NodeValidator.cpp", "jsFunction_validateOneOf", 0);
+
